lab-eval/requirements.txt
azure-ai-projects
azure-ai-inference[prompts]
azure-identity
azure-search-documents
pandas
python-dotenv
opentelemetry-api
azure-monitor-opentelemetry
azure-ai-evaluation[remote]
